feat: Gestures support

+++ engine/src/input/native-input-handling.h
@@ -40,12 +40,12 @@ namespace input
       {
          static inline void OnKeyStateChanged(uint32_t key, uint32_t sc, uint32_t action, uint32_t mods)
          {
-            g_CurrentFrameEventStates[static_cast<uint8_t>(mapping::g_GlfwInputEventMap[key])] = (action != GLFW_RELEASE) ? true : false;
+            g_CurrentFrameEventStates[static_cast<size_t>(mapping::g_GlfwInputEventMap[key])] = (action != GLFW_RELEASE) ? true : false;
          }
 
          static inline void OnMouseButtonStateChanged(uint32_t button, uint32_t action, uint32_t mods)
          {
-            g_CurrentFrameEventStates[static_cast<uint8_t>(mapping::g_GlfwInputEventMap[1 + button + (uint8_t)(input::InputEvent::LastKeyboardKey)])] = (action != GLFW_RELEASE) ? true : false;
+            g_CurrentFrameEventStates[static_cast<size_t>(mapping::g_GlfwInputEventMap[1 + button + (size_t)(input::InputEvent::LastKeyboardKey)])] = (action != GLFW_RELEASE) ? true : false;
          }
       };
